image_processing_service/requirements.txt
Flask
requests
Pillow
google-generativeai
transformers